bot/requirements.txt
python_telegram_bot>=20.2
requests>=2.28.2
